backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
ast_decompiler==0.8.0
click==8.2.1
colorama==0.4.6
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
joblib==1.5.1
nltk==3.9.1
numpy==2.3.2
pandas==2.3.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic-bin==0.4.14
python-multipart==0.0.20
pytz==2025.2
regex==2025.7.34
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sqlparse==0.5.3
textdistance==4.6.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
zipfile36==0.1.3
